+++ Changelog.md
@@ -6,6 +6,8 @@ Bug Fixes:
 * Ensure custom error codes are returned from bisect runs. (Jon Rowe, #2732)
 * Ensure `RSpec::Core::Configuration` predicate config methods return booleans.
   (Marc-André Lafortune, #2736)
+* Prevent `rspec --bisect` from generating zombie processes while executing
+  bisect runs. (Benoit Tigeot, Jon Rowe, #2739)
 
 ### 3.9.2 / 2020-05-02
 [Full Changelog](http://github.com/rspec/rspec-core/compare/v3.9.1...v3.9.2)